Fix: balance, amount 타입 수정

+++ src/main/java/team/hanaro/hanamate/domain/MyWallet/Dto/ResponseDto.java
@@ -18,8 +18,8 @@ public class ResponseDto {
     public static class MyWallet {
         private Long walletId;
         private Boolean walletType;
-        private Long balance;
-        private Long targetAmount;
+        private Integer balance;
+        private Integer targetAmount;
 
         public MyWallet(Wallets wallets) {
             walletId = wallets.getWalletId();
